refactor: CharacterRepository
