+++ fourtooncookie/src/main/java/com/startingblue/fourtooncookie/character/CharacterRepository.java
@@ -1,6 +1,8 @@
-package com.startingblue.fourtooncookie.character.domain;
+package com.startingblue.fourtooncookie.character;
 
 import com.startingblue.fourtooncookie.artwork.domain.Artwork;
+import com.startingblue.fourtooncookie.character.domain.Character;
+import com.startingblue.fourtooncookie.character.domain.CharacterVisionType;
 import com.startingblue.fourtooncookie.global.domain.PaymentType;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
