+++ fourtooncookie/src/main/java/com/startingblue/fourtooncookie/character/CharacterService.java
@@ -3,7 +3,6 @@ package com.startingblue.fourtooncookie.character;
 import com.startingblue.fourtooncookie.artwork.domain.Artwork;
 import com.startingblue.fourtooncookie.artwork.service.ArtworkService;
 import com.startingblue.fourtooncookie.character.domain.Character;
-import com.startingblue.fourtooncookie.character.domain.CharacterRepository;
 import com.startingblue.fourtooncookie.character.domain.CharacterVisionType;
 import com.startingblue.fourtooncookie.global.domain.PaymentType;
 import com.startingblue.fourtooncookie.character.dto.request.CharacterSaveRequest;
