+++ fourtooncookie/src/test/java/com/startingblue/fourtooncookie/diary/domain/DiaryRepositoryTest.java
@@ -3,7 +3,7 @@ package com.startingblue.fourtooncookie.diary.domain;
 import com.startingblue.fourtooncookie.artwork.domain.Artwork;
 import com.startingblue.fourtooncookie.artwork.ArtworkRepository;
 import com.startingblue.fourtooncookie.character.domain.Character;
-import com.startingblue.fourtooncookie.character.domain.CharacterRepository;
+import com.startingblue.fourtooncookie.character.CharacterRepository;
 import com.startingblue.fourtooncookie.character.domain.CharacterVisionType;
 import com.startingblue.fourtooncookie.diary.DiaryRepository;
 import com.startingblue.fourtooncookie.global.domain.PaymentType;
